% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoCOPoint.R
\name{GenoCOPoint}
\alias{GenoCOPoint}
\title{Detect CO Points}
\usage{
GenoCOPoint(
  geno_file,
  truth_file,
  output_pdf = "CO_points.pdf",
  output_txt = "COresult.txt",
  window_size = 5,
  smooth_threshold = 0.6,
  min_support = 7,
  consistency = 0.85
)
}
\arguments{
\item{geno_file}{The path to the genotype file with three columns of information 
(name of the chromosome where the SNP is located,
absolute position of the SNP,
code of the genotype corresponding to the SNP 0/1,
read counts of ref genotype,
read counts of alt genotype) .}

\item{truth_file}{The path to the truth file with CO intervals.
The file should contain the chromosome name, start and end positions of the CO intervals.}

\item{output_pdf}{Path to the output PDF file for visualization.}

\item{output_txt}{Path to the output text file for results.}

\item{window_size}{Size of the sliding window for smoothing.}

\item{smooth_threshold}{Threshold for smoothing.}

\item{min_support}{Minimum number of consecutive SNPs for noise filtering.}

\item{consistency}{Consistency ratio for filtering.}
}
\value{
A list where each element contains the CO positions for a chromosome.
}
\description{
This function reads a genotype file, processes the data, detects CO points and generates a PDF visualization and a result file..
}
\examples{
# Example usage:
# Create a temporary input file
geno_file <- "GenoCOPoint/examples/genofiles/AAACGCTGTGATATAG_smt_genotypes_withname_reads.txt"
truth_file <- "GenoCOPoint/examples/truthfiles/AAACGCTGTGATATAG_allele_cnts_at_markers_sorted_co_pred.txt"
# Run the function
result <- GenoCOPoint(geno_file,truth_file, output_pdf = "example_output.pdf", output_txt = "example_output.txt", window_size = 5, smooth_threshold = 0.6, min_support = 7, consistency = 0.85)
print(result)

# Clean up
unlink(geno_file)
unlink(truth_file)
unlink("example_output.pdf")
unlink("example_output.txt")
}
